Move Router to components/ too
