+++ src/index.js
@@ -4,8 +4,8 @@ import React from 'react';
 import { render } from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
 
-import Router from './routes';
 import config from './config';
+import Router from './components/Router';
 
 const mountNode = document.getElementById('app');
 
@@ -24,5 +24,5 @@ renderApp();
 
 // Hot Module Replacement API
 if (module.hot) {
-  module.hot.accept('./routes', () => renderApp());
+  module.hot.accept('./components/Router', () => renderApp());
 }
